+++ src/new/linux_uapi/linux/mod.rs
@@ -1,0 +1,4 @@
+//! The `linux` directory within `include/uapi` in the Linux source tree.
+
+pub(crate) mod can;
+pub use can::*;
