+++ src/new/linux_uapi/mod.rs
@@ -1,0 +1,4 @@
+//! This directory maps to `include/uapi` in the Linux source tree.
+
+pub(crate) mod linux;
+pub use linux::*;
